requirements.txt
certifi==2019.6.16
cffi==1.12.3
easydict==1.9
numpy==1.17.0
olefile==0.46
Pillow==6.0.0
protobuf==3.9.1
pycparser==2.19
six==1.12.0
tensorboardX==1.8
